feat: merge backend connections + networks -> WifiNetwork.
Adds NMWirelessManager which merges identical NMWirelessNetworks (groups
of APs with the same SSID) and NMConnectionAdapters (saved connection
settings to wifi networks), and creates WifiNetworks on the frontend.
The manager takes a similar approach to plasma-nm and nm-applet. Adds
utilities to determine a connections security type and a networks best
choice for security, in sync with nm_utils_security_valid from nmcli.